// typescript-projects/utility-backups/src20250528BeforeEbillsTokenFromEnv/interface/types.ts
export type TEbillsCableTvPaymentData = {
  provider: ProviderId;
  cardNumber: string;
  merchant: string;
  variationId: string;
};